requirements.txt
pytest
pylint
jupyter
pytest-cov
pandas
nbval
click
flask
requests
google-cloud-storage
google-cloud-bigquery
bs4
flasgger
google-api-python-client
locust
pyarrow
matplotlib